Fix broken realm sorting
This fixes bugs #2520 and #2484
The cause of the bugs is the fact that calling `.sorted` method
on the realm resultset produces another resultset which is sorted
when iterated through `Iterator<T>` interface, but does absolutely
nothing to way how resultset object is serialised (printed out
as JSON).
So when we sorted, printed to json, parsed back through transit
and expected the order to be retained, it was not, and it was
failing quite randomly.
To be precise, it failed in a case where we sorted the resultset
according to some property and picked the top element from the result.
As the top element picking was done AFTER the sorted realm resultset
was converted into cljs datastructure, it was sometimes failing.

This fix just ensures that we select the single element from js
resultset (where sorting info is still retained) and only convert
that single element to cljs (it's also much more efficient).