Denormalize message on replying
Before we stored only the message-id and had a subscription pulling the
message from the database when replying to the message.
This broke once we implemented offloading of messages, as the message
might not be in the database anymore.
This commit fixes the issue by storing the full message in the database,
so in the event of it being offloaded it is still shown.

Signed-off-by: Andrea Maria Piana <andrea.maria.piana@gmail.com>